requirements.txt
Flask==3.0.3
gunicorn 
python-dotenv==1.0.1
pandas==2.2.2
SQLAlchemy==2.0.30
matplotlib==3.9.0
langchain==0.3.27
langchain-community==0.3.27
langchain-openai==0.3.28
langchain-experimental==0.3.1
openai>=1.14.0
tabulate==0.9.0
tiktoken>=0.6.0         
FLASK_SECRET_KEY=f5a712fe70c4498c1e80b9a3df4e8e746841d2bf51a4c9a47b11a2e1f3f0e654

